requirements.txt
weave
openai>=1.0.0
python-dotenv
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
psycopg2
pgvector